+++ src/main/java/com/github/unaimillan/rars/riscv/instructions/SLLIW.java
@@ -14,6 +14,7 @@ public class SLLIW extends BasicInstruction {
     public void simulate(ProgramStatement statement) {
         // Copy from SLLI
         int[] operands = statement.getOperands();
-        RegisterFile.updateRegister(operands[0], (long) RegisterFile.getValue(operands[1]) << operands[2]);
+        // Shift implicitly casts to (long) type, but it should be here to sign-extend the value
+        RegisterFile.updateRegister(operands[0], RegisterFile.getValue(operands[1]) << operands[2]);
     }
 }
